+++ formulaire/frontend/src/core/models/form/utils.ts
@@ -2,7 +2,8 @@ import { FormPropField } from "~/containers/FormPropModal/enums";
 import { IFormPropInputValueState } from "~/containers/FormPropModal/types";
 import { IFormPayload, IForm } from "./types";
 import { IDistribution } from "../distribution/types";
-import { getLatestDistribution, getNbFinishedDistrib } from "../distribution/utils";
+import { getFirstDistribution, getLatestDistribution, getNbFinishedDistrib } from "../distribution/utils";
+import { DistributionStatus } from "../distribution/enums";
 
 export const buildFormPayload = (
   formPropValue: IFormPropInputValueState,
@@ -69,12 +70,11 @@ export const getFormDistributions = (form: IForm, distributions: IDistribution[]
 };
 
 export const isFormFilled = (form: IForm, distributions: IDistribution[]): boolean => {
+  const formDistributions = getFormDistributions(form, distributions);
   if (form.multiple) {
-    const latesteDistrib = getLatestDistribution(distributions);
-    return !!latesteDistrib.dateResponse;
-  } else {
-    return getNbFinishedDistrib(distributions) > 0;
+    return getFirstDistribution(formDistributions).status === DistributionStatus.FINISHED;
   }
+  return getNbFinishedDistrib(distributions) > 0;
 };
 
 export const getFormStatusText = (
@@ -83,11 +83,12 @@ export const getFormStatusText = (
   formatDateWithTime: (date: string | Date | undefined, i18nTextKey: string) => string,
   t: (key: string) => string,
 ): string => {
+  const formDistributions = getFormDistributions(form, distributions);
   if (form.multiple) {
-    return `${t("formulaire.responses.count")} : ${getNbFinishedDistrib(distributions).toString()}`;
+    return `${t("formulaire.responses.count")} : ${getNbFinishedDistrib(formDistributions).toString()}`;
   } else {
-    if (getNbFinishedDistrib(distributions) > 0) {
-      const latestDistrib = getLatestDistribution(distributions);
+    if (getNbFinishedDistrib(formDistributions) > 0) {
+      const latestDistrib = getLatestDistribution(formDistributions);
       if (latestDistrib.dateResponse) {
         return formatDateWithTime(latestDistrib.dateResponse, "formulaire.responded.date");
       }
